[core-http] mark de-serialized properties as handled
Currently we copy additional properties into the result if they are not marked
as handled. In cases where `xmlName` is different from `seralizedName`, we end
up copying unwanted additional property that has been de-serialized already
because the marked flag is set for `serializedName`.

This PR marks an XML property as handled after it is de-serialized.