Cherry pick abort-controller changes for abort-controller and core-http
[abort-controller] Remove polyfills for `AbortController` and `AbortSignal` (#27921)

`@azure/abort-controller`

Fixes #27018

As `AbortController` and `AbortSignal` now exist in all supported
runtimes, we do not need to have our own userland implementation of
them.

This PR creates a new major version of the `abort-controller` package
that only has `AbortSignalLike` and `AbortError` and updates core +
test-utils to use the new version.