[engsys] upgrade dev dependency `eslint-plugin-markdown` to ^5.0.0
The breaking change that impacts us is that now `typescript` and `javascript`
code blocks are maped to virtual files with `.ts` or `.js`, instead of
`.typescript` or `.javascript` previously in v4.

This PR reacts to the change:

- Add an overrides entry for `**/*.md/*.ts` because they now matches the files
pattern of our first override. Previously they are not linted because of the old
`.typescript` virtual file extension.

- remove `--ext` option passed to `eslint`. According to the doc specified file
extensions via this option will take precedence over the `overrides` entries in
the config and break the `**/*.md/*.ts` overrides. Two ignore patterns for
`.json` are added since they matched the "*.json" pattern with `--ext` removed.

***NO_CI***